+++ app/src/main/java/org/vault/CryptoUtils.java
@@ -1,4 +1,4 @@
-package org.YAPM;
+package org.vault;
 
 import javax.crypto.Cipher;
 import javax.crypto.SecretKey;
