+++ app/src/main/java/org/vault/VaultStatus.java
@@ -1,4 +1,4 @@
-package org.YAPM;
+package org.vault;
 
 public enum VaultStatus {
   DBConnectionSuccess,
